requirements.txt
google-play-scraper==1.2.4
pandas>=2.0.0
python-dotenv>=1.0.0
snowflake-connector-python>=3.10
snowflake-sqlalchemy>=1.5
pyarrow>=16.0